ML_Training/flask_app/requirements.txt
Flask==2.2.3
keras==3.7.0
tensorflow==2.18.0
flask-cors==5.0.0
pillow==9.5.0
scikit-learn==1.6.0
werkzeug==2.2.3
gunicorn==23.0.0
matplotlib==3.8.0
